Merge windowed statistics implementation (Phase 3.2)

Add local windowed statistics (mean, mean-square, variance, RMS)
using integral images for O(1) per-pixel computation.

Functions added:
- mean_square_accum: f64-precision integral image of squared values
- windowed_mean: local mean with optional normalization
- windowed_mean_square: local mean of squared values
- windowed_variance: variance and RMS from precomputed statistics
- windowed_stats: high-level convenience function

Copilot review addressed:
- Clamp negative variance to zero before sqrt (fixes NaN propagation)
- Add +0.5 rounding to windowed_mean normalization
- Improve dimension check error messages
- Clarify f64 precision rationale (rounding vs overflow)
- Simplify test comment derivations

18 unit tests covering correctness and error cases.

+++ crates/leptonica-filter/src/block_conv.rs
@@ -257,7 +257,7 @@ mod tests {
         let mut pm = pix.try_into_mut().unwrap();
         for y in 0..h {
             for x in 0..w {
-                let val = ((x + y * w) % 256) as u32;
+                let val = (x + y * w) % 256;
                 pm.set_pixel_unchecked(x, y, val);
             }
         }
